Add some basic tests. Update httpclient -> httpclient wrapperin DI
